FAT-130 Refactor API client
