--- acquisitions/src/test/java/com/gurock/testrail/APIException.java
@@ -1,20 +1,0 @@
-/**
- * TestRail API binding for Java (API v2, available since TestRail 3.0)
- *
- * Learn more:
- *
- * http://docs.gurock.com/testrail-api2/start
- * http://docs.gurock.com/testrail-api2/accessing
- *
- * Copyright Gurock Software GmbH. See license.md for details.
- */
-
-package com.gurock.testrail;
- 
-public class APIException extends Exception
-{
-	public APIException(String message)
-	{
-		super(message);
-	}
-}
