+++ acquisitions/src/test/java/org/folio/testrail/api/APIException.java
@@ -1,0 +1,8 @@
+package org.folio.testrail.api;
+
+public class APIException extends Exception {
+
+  public APIException(String message) {
+    super(message);
+  }
+}
