+++ acquisitions/src/test/java/org/folio/testrail/api/TestsApiEntry.java
@@ -1,7 +1,5 @@
-package com.gurock.testrail.api;
+package org.folio.testrail.api;
 
-import com.gurock.testrail.APIClient;
-import com.gurock.testrail.APIException;
 import java.io.IOException;
 import org.json.simple.JSONArray;
